acpi, movablemem_map: make whatever nodes the kernel resides in un-hotpluggable.
There could be several memory ranges in the node in which the kernel
resides.  When using movablemem_map=acpi, we may skip one range that have
memory reserved by memblock.  But if it is too small, then the kernel will
fail to boot.  So, make the whole node which the kernel resides in
un-hotpluggable.  Then the kernel has enough memory to use.

Signed-off-by: Tang Chen <tangchen@cn.fujitsu.com>
Reported-by: H Peter Anvin <hpa@zytor.com>
Cc: KOSAKI Motohiro <kosaki.motohiro@jp.fujitsu.com>
Cc: Jiang Liu <jiang.liu@huawei.com>
Cc: Jianguo Wu <wujianguo@huawei.com>
Cc: Kamezawa Hiroyuki <kamezawa.hiroyu@jp.fujitsu.com>
Cc: Lai Jiangshan <laijs@cn.fujitsu.com>
Cc: Wu Jianguo <wujianguo@huawei.com>
Cc: Yasuaki Ishimatsu <isimatu.yasuaki@jp.fujitsu.com>
Cc: Ingo Molnar <mingo@elte.hu>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Len Brown <lenb@kernel.org>
Cc: "Brown, Len" <len.brown@intel.com>
Cc: Will Huck <will.huckk@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>